requirements_ocr.txt
argh==0.26.2
bitstring==3.1.5
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorlog==3.1.4
ConfigArgParse==0.13.0
cycler==0.10.0
Flask==1.0.2
Flask-Caching==1.4.0
future==0.17.1
gpxdata==1.2.1
idna==2.7
ImageHash==4.0
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mysql-connector==2.1.6
mysqlclient==1.3.14
numpy==1.15.4
opencv-python==3.4.4.19
pathtools==0.1.2
Pillow==5.3.0
pyparsing==2.3.0
pytesseract==0.2.5
python-dateutil==2.7.5
PyWavelets==1.0.1
PyYAML==3.13
requests==2.20.1
s2sphere==0.2.5
scipy==1.1.0
six==1.11.0
urllib3==1.24.1
watchdog==0.9.0
websockets==7.0
Werkzeug==0.14.1
